requirements.txt
python-dotenv==1.0.1
openai
chromadb==0.4.13